:recycle: #29 - createManyMapper
PrismaHelper 클래스 생성 및 createManyMapper 메서드 변경
