+++ src/modules/core/database/prisma/prisma.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { PrismaService } from '@src/modules/core/database/prisma/prisma.service';
+import { PrismaHelper } from './prisma.helper';
 
 @Module({
-  providers: [PrismaService],
-  exports: [PrismaService],
+  providers: [PrismaService, PrismaHelper],
+  exports: [PrismaService, PrismaHelper],
 })
 export class PrismaModule {}
